backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.27.0
python-multipart==0.0.9
stripe==8.4.0
sse-starlette==2.0.0
python-dotenv==1.0.1
asyncio==3.4.3
aiosqlite==0.19.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.3
email-validator==2.1.0
